ignore symbols without addresses for symbolization
There are various types of symbols in object symbols tables whose
"value" isn't a meaningful address and thus shouldn't be used for
symbolizing an address. For example, absolute symbols and TLS symbols.
Currently, these often cause small integer constants in assembly to
get symbolized as nonsense symbols.

Plumb whether or not a symbol's value is an address through to the
generic symbol table so these can be ignored for symbolization.