handle relocations and display them in hex view
For a seemingly small change, this basically required a complete
redesign of the helper packages.

This adds a generalized relocation representation. Data loaded from an
object is now always represented by a byte slice *and* a sequence of
relocations.

Since relocations often target the dynamic symbol table, the ELF
loader now reads in both symbol tables and combines them.

This updates the hex view to display relocations at the positions
where they apply.

In the process, this fixes a bunch of other relocated things:

- It introduces a global "symbol ID" taken from all combined symbol
  tables. This lets the generic relocation representation reference a
  symbol using a single ID, which can then easily be passed back in to
  an object file's methods to get data on that symbol.

- Sequences of symbols and relocs are now both represented by sequence
  interfaces that allow for lazy decoding into the common
  representation.

- Symbol size synthesis is now much more sophisticated. It handles
  symbols across different sections correctly, multiple symbols at the
  same base address, and symbols without addresses.

- This fixes reading section data from SHT_NOBITS sections.

- symtab.Table no longer mutates the order of the symbols slice. This
  is important for using symbol IDs.

- symtab.Table now returns symbol IDs. The caller can turn this into a
  symbol.