cmd/objbrowse: rebuild views on entity change
Currently, because the list of View elements is keyed only by the view
ID, when the user switches to a different entity, React will carefully
update the existing view to adapt it from the old entity to the new
entity. This is almost certainly more work than just creating a new
view from scratch, but also means that the view currently keeps its
scroll position.

Fix this by making the entity part of the view element key, so that
when the entity changes, React discards the old views entirely and
builds new ones.