cmd/objbrowse: much improved element scrolling
Currently, we use scrollIntoView to scroll to the current selection,
but this doesn't work on display:none elements, so it only scrolls the
currently visible view. If we fix that, then we run into a Chrome bug
wherein starting a second scrollIntoView will cancel the first, so we
wind up only scrolling one view (though a different only one view).

Fix this by replacing display:none with visibility:hidden for the
views so they can be scrolled and using scrollTo, which doesn't have
the cancellation problem of scrollIntoView. This also gives us the
opportunity to customize scroll behavior a little more and to scroll
all selected elements into view rather than just the first one.