cmd/objbrowse: rationalize current view state
Currently, the current view state is kept in the App and updating the
selection resets the current view and the valid views. But this leads
to a bug where clicking a link to a different entity in asmview
changes the selected entity, doesn't reset the valid views.

But this is really just the tip of the ice berg of the issue. The
current view shouldn't be owned by App at all, and it's really awkward
for App to reset it when the entity changes because deriving one state
change from another in React is difficult. And the valid views isn't
state at all, but just a computation from the current entity.

Hence, move the current view down into EntityPanel's state. Compute
valid views directly from the selection in App and pass it to
EntityPanel (we could alternatively compute it there). Give
EntityPanel a key so it gets rebuilt when the entity changes and
recomputes the default view. This also simplifies some messy keys
inside EntityPanel. Finally, this reveals a bug where clicking a link
to another entity in asmview both selects that entity and *then*
selects the row, which re-selects the current entity with a different
range. Fix that by stopping event propagation. Whew.