cmd/objbrowse: don't re-render SymPanel on range changes
We need to memoize the setEntity callback so it's not new on every App
re-render, and tweak asmview so it uses the current entity object for
self-links. Other range changes use onSelectRange, which already keeps
the entity object the same.

While we're here, now that I understand React.useCallback, wrap
another callback.