cmd/objbrowse: support multiple view panes
This adds buttons to add view columns and close view columns. All of
the views remain in sync.

Currently, this always reverts to a single column when switching
entities. It's possible I actually want to persist the panes I have
and their currently selected view, though this is difficult if you
navigate to an entity that has a different set of views.

I should at least be saving the view state in the browser history, now
that it's much more complex and thus annoying to recreate when you hit
"back". But I don't have the framework in place to do that yet.