cmd/objbrowse: rework window history to use state objects
This totally redoes how we handle window history. Now it uses window
history state objects, which are wrapped around React state and
carefully hooked into the component tree. This lets us track a lot
more in the state, including the entity panel layout and selected
views. It also lets us distinguish between when we want to push an
entirely new state on the history stack, or when we want to just
replace the state so that if the user navigates away they'll come back
to the current state.