cmd/objbrowse: add FetchJSON higher-order component
This provides a higher-order component that wraps useFetchJSON and
switches all of the views to use this. In general this is a bit easier
to use: it forces the caller to use a second component for once the
data is loaded, which means that component is more focused and can use
hooks much more naturally. Previously, if a component wasn't already
split into two, it had to invoke all hooks before the early exit on
fetch.pending. This constraint had already forced me to move to a
two-level component approach in two places. FetchJSON makes this more
natural.