cmd/objbrowse: fix scrolling in new views
Currently, the new views created when a new entity is selected don't
scroll to the selected region. This happens because the scrolling is
attempted before the views have finished asynchronously loading their
data.

Fix this by putting scrolling behavior into a new higher-order
component that's only created once the view's data is loaded.